ABOUT THE TASK
ATM transactions are represented by ATM software. In this case, the user must choose an option from the possibilities shown on the screen. The choices pertain to taking money out, investing cash in, checking the balance, and leaving.

We just obtain the user's requested withdrawal amount, deduct it from the balance total, and publish a successful message.

To deposit the funds, we only request the user's deposit amount, add it to the balance overall, and print the successful message.

We print the user's entire balance to check the balance.

The exit(0) method is used to quit the active Transaction mode and take the user back to the login screen or home page.